requirements.txt
# Core data science libraries
pandas>=1.5.0
numpy>=1.21.0
matplotlib>=3.5.0
seaborn>=0.11.0

# Data sources
yfinance>=0.2.0
quandl>=3.7.0
tinkoff-investments>=0.2.0

# Statistical analysis
scipy>=1.9.0
statsmodels>=0.13.0

# Backtesting and financial analysis
# Replaced empyrical with pandas-ta for technical analysis
pandas-ta>=0.3.14b0
# Replaced pyfolio with pyfolio-reloaded (maintained fork)
pyfolio-reloaded>=0.9.2

# Jupyter notebooks
jupyter>=1.0.0
ipykernel>=6.0.0

# Configuration and utilities
pyyaml>=6.0
python-dotenv>=0.19.0

# Testing
pytest>=7.0.0
pytest-cov>=4.0.0

# Code quality
black>=22.0.0
flake8>=5.0.0 